+++ supreme/verifier/src/commonTest/kotlin/main.kt
@@ -1,7 +1,6 @@
 package at.asitplus.attestation.supreme
 
-import at.asitplus.attestation.IOSAttestationConfiguration
-import at.asitplus.attestation.Warden
+import at.asitplus.attestation.IosAttestationConfiguration
 import at.asitplus.attestation.android.AndroidAttestationConfiguration
 import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.asn1.Asn1String
@@ -25,7 +24,6 @@ import io.ktor.server.plugins.contentnegotiation.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
-import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.json.Json
 import kotlin.random.Random
 import kotlin.time.Clock
@@ -49,32 +47,36 @@ class TestEnv : FreeSpec({
         var running = true
 
         val attestationValidator = AttestationValidator(
-            Warden(
-                AndroidAttestationConfiguration.Builder(
-                    AndroidAttestationConfiguration.AppData(
-                        "at.asitplus.attestation.supreme.client.test", //automated tests
-                        listOf(
-                            "a3e55ba9457de2900fe86303a5d556c496b691afff2c0dd50488bed3e400cc6b".hexToByteArray(
-                                HexFormat.Default
-                            )
+            AndroidAttestationConfiguration.Builder(
+                AndroidAttestationConfiguration.AppData(
+                    "at.asitplus.attestation.supreme.client.test", //automated tests
+                    listOf(
+                        "a3e55ba9457de2900fe86303a5d556c496b691afff2c0dd50488bed3e400cc6b".hexToByteArray(
+                            HexFormat.Default
                         )
                     )
-                ).enableSoftwareAttestation().disableHardwareAttestation().addSoftwareAttestationTrustAnchor(
-                    CryptoPublicKey.decodeFromPem("-----BEGIN PUBLIC KEY-----\n" +
+                )
+            ).enableSoftwareAttestation().disableHardwareAttestation().addSoftwareAttestationTrustAnchor(
+                CryptoPublicKey.decodeFromPem(
+                    "-----BEGIN PUBLIC KEY-----\n" +
                             "MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE9+hz7A0vjTx6w2x7E6wW8Cy3MlJY\n" +
                             "+E3HadGEUI8McOFz3VytQgylZWfT+LUKDjTq3CBffGbo1GeBH+leQlFoaw==\n" +
-                            "-----END PUBLIC KEY-----").getOrThrow().toJcaPublicKey().getOrThrow()
-                ).build(),
-                IOSAttestationConfiguration(
-                    IOSAttestationConfiguration.AppData(
-                        "9CYHJNG644",
-                        "at.asitplus.signumtest.iosApp", //to test with real app from ios
-                        sandbox = true
-                    )
-                ), Clock.System, verificationTimeOffset = 3.minutes
+                            "-----END PUBLIC KEY-----"
+                ).getOrThrow().toJcaPublicKey().getOrThrow()
+            ).build(),
+            IosAttestationConfiguration(
+                IosAttestationConfiguration.AppData(
+                    "9CYHJNG644",
+                    "at.asitplus.signumtest.iosApp", //to test with real app from ios
+                    sandbox = true
+                )
             ),
             attestationProofOID = PROOF_OID,
-        ) { it.contentEquals(NONCE) }
+            Clock.System, verificationTimeOffset = 3.minutes,
+        ) {
+            if (it.contentEquals(NONCE)) ChallengeValidationResult.Success
+            else ChallengeValidationResult.Failure(null)
+        }
 
 
         val server = embeddedServer(Netty, port = 8080) {
@@ -85,52 +87,59 @@ class TestEnv : FreeSpec({
             routing {
                 get("/shutdown") {
                     call.respondText("Bye!")
-                    running=false
+                    running = false
                 }
 
                 get(ENDPOINT_CHALLENGE) {
 
                     println("Issuing Challenge")
-                    call.respondText(Json.encodeToString(attestationValidator.issueChallenge(NONCE, 10.minutes, ENDPOINT_ATTEST, timeOffset = -5.minutes)), contentType = ContentType.Application.Json)
+                    call.respondText(
+                        Json.encodeToString(
+                            attestationValidator.issueChallenge(
+                                NONCE,
+                                10.minutes,
+                                ENDPOINT_ATTEST,
+                                timeOffset = -5.minutes
+                            )
+                        ), contentType = ContentType.Application.Json
+                    )
                 }
                 post(PATH_ATTEST) {
                     val src = call.receive<ByteArray>()
-
                     val resp =
                         attestationValidator.verifyKeyAttestation(Pkcs10CertificationRequest.decodeFromDer(src)) { csr ->
                             Signer.Ephemeral {
                                 ec { }
                             }.getOrThrow().let { signer ->
-                                runBlocking {
-                                    signer.sign(
-                                        TbsCertificate(
-                                            serialNumber = Random.nextBytes(32),
-                                            publicKey = signer.publicKey,
-                                            signatureAlgorithm = signer.signatureAlgorithm.toX509SignatureAlgorithm()
-                                                .getOrThrow(),
-                                            validFrom = Asn1Time(Clock.System.now()),
-                                            validUntil = Asn1Time(Clock.System.now() + 10.days),
-                                            issuerName = listOf(
-                                                RelativeDistinguishedName(
-                                                    AttributeTypeAndValue.CommonName(
-                                                        Asn1String.UTF8(
-                                                            "WARDEN Supreme"
-                                                        )
+                                signer.sign(
+                                    TbsCertificate(
+                                        serialNumber = Random.nextBytes(32),
+                                        publicKey = signer.publicKey,
+                                        signatureAlgorithm = signer.signatureAlgorithm.toX509SignatureAlgorithm()
+                                            .getOrThrow(),
+                                        validFrom = Asn1Time(Clock.System.now()),
+                                        validUntil = Asn1Time(Clock.System.now() + 10.days),
+                                        issuerName = listOf(
+                                            RelativeDistinguishedName(
+                                                AttributeTypeAndValue.CommonName(
+                                                    Asn1String.UTF8(
+                                                        "WARDEN Supreme"
                                                     )
                                                 )
-                                            ),
-                                            subjectName = csr.tbsCsr.subjectName,
-                                        )
-                                    ).map { listOf(it) }
-                                }
+                                            )
+                                        ),
+                                        subjectName = csr.tbsCsr.subjectName,
+                                    )
+                                ).map { listOf(it) }
                             }
                         }
-
                     call.respondText(Json.encodeToString(resp), contentType = ContentType.Application.Json)
                 }
             }
         }.start(wait = false)
         println("KTOR server started!")
-        while(running) {Thread.sleep(500)}
+        while (running) {
+            Thread.sleep(500)
+        }
     }
 })